trend-analysis-platform/backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP client and external APIs
httpx==0.25.2
aiohttp==3.9.1

# Data processing
pandas==2.1.4
numpy==1.25.2
pydantic==2.5.0
pydantic-settings==2.1.0

# Background tasks
celery==5.3.4
redis==5.0.1

# Monitoring and logging
structlog==23.2.0
prometheus-client==0.19.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Environment management
python-dotenv==1.0.0